requirements.txt
selenium
pandas
numpy
seaborn
matplotlib
streamlit
webdriver-manager
plotly
mysql-connector-python